+++ src/main/java/eu/dissco/core/handlemanager/domain/requests/PatchRequestData.java
@@ -3,7 +3,7 @@ package eu.dissco.core.handlemanager.domain.requests;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import com.fasterxml.jackson.databind.JsonNode;
-import eu.dissco.core.handlemanager.domain.ObjectType;
+import eu.dissco.core.handlemanager.domain.requests.attributes.ObjectType;
 
 public record PatchRequestData(
     @JsonProperty(required = true)
