+++ src/main/java/eu/dissco/core/handlemanager/domain/requests/attributes/ObjectType.java
@@ -1,4 +1,4 @@
-package eu.dissco.core.handlemanager.domain;
+package eu.dissco.core.handlemanager.domain.requests.attributes;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
